[ADD] event_product: extract product from sales related module *
In order to use the link between products and events in the
point_of_sale without having to install the sale_* module suite, this
commit extracts this link from the event_booth_sale, event_sale,
website_event_sale modules.
