+++ addons/event_booth_sale/data/product_data.xml
@@ -6,7 +6,7 @@
         <field name="purchase_ok" eval="False"/>
         <field name="list_price">100.0</field>
         <field name="description_sale" eval="False"/>
-        <field name="categ_id" ref="event_sale.product_category_events"/>
+        <field name="categ_id" ref="event_product.product_category_events"/>
         <field name="invoice_policy">order</field>
         <field name="type">service</field>
         <field name="service_tracking">event_booth</field>
