+++ addons/event_product/data/event_demo.xml
@@ -1,40 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-
     <record id="event.event_0_ticket_0" model="event.event.ticket">
         <field name="product_id" ref="product_product_event"/>
         <field name="price">0</field>
     </record>
     <record id="event.event_0_ticket_1" model="event.event.ticket">
-        <field name="product_id" ref="product_product_event_standard"/>
+        <field name="product_id" ref="event_product.product_product_event_standard"/>
         <field name="price">1000.0</field>
     </record>
     <record id="event.event_0_ticket_2" model="event.event.ticket">
-        <field name="product_id" ref="product_product_event_vip"/>
+        <field name="product_id" ref="event_product.product_product_event_vip"/>
         <field name="price">1500.0</field>
     </record>
 
     <record id="event.event_2_ticket_1" model="event.event.ticket">
-        <field name="product_id" ref="product_product_event_standard"/>
+        <field name="product_id" ref="event_product.product_product_event_standard"/>
         <field name="price">1000.0</field>
     </record>
     <record id="event.event_2_ticket_2" model="event.event.ticket">
-        <field name="product_id" ref="product_product_event_vip"/>
+        <field name="product_id" ref="event_product.product_product_event_vip"/>
         <field name="price">1500.0</field>
     </record>
 
     <record id="event.event_4_ticket_0" model="event.event.ticket">
-        <field name="product_id" ref="product_product_event_standard"/>
+        <field name="product_id" ref="event_product.product_product_event_standard"/>
         <field name="price">99.0</field>
     </record>
 
     <record id="event.event_7_ticket_1" model="event.event.ticket">
-        <field name="product_id" ref="product_product_event_standard"/>
+        <field name="product_id" ref="event_product.product_product_event_standard"/>
         <field name="price">0.0</field>
     </record>
     <record id="event.event_7_ticket_2" model="event.event.ticket">
-        <field name="product_id" ref="product_product_event_vip"/>
+        <field name="product_id" ref="event_product.product_product_event_vip"/>
         <field name="price">0.0</field>
     </record>
-
 </odoo>
