+++ addons/event_product/models/__init__.py
@@ -1,0 +1,5 @@
+from . import event_event
+from . import event_event_ticket
+from . import event_type_ticket
+from . import product_product
+from . import product_template
