+++ addons/event_product/models/event_event.py
@@ -1,0 +1,9 @@
+from odoo import fields, models
+
+
+class Event(models.Model):
+    _inherit = 'event.event'
+
+    currency_id = fields.Many2one(
+        'res.currency', string='Currency',
+        related='company_id.currency_id', readonly=True)
