+++ addons/event_product/models/product_product.py
