+++ addons/event_product/models/product_template.py
@@ -1,0 +1,9 @@
+from odoo import fields, models
+
+
+class ProductTemplate(models.Model):
+    _inherit = 'product.template'
+
+    service_tracking = fields.Selection(selection_add=[
+        ('event', 'Event Registration'),
+    ], ondelete={'event': 'set default'})
