+++ addons/event_product/views/event_ticket_views.xml
@@ -3,7 +3,7 @@
 
     <!-- EVENT.TYPE.TICKET -->
     <record id="event_type_ticket_view_tree_from_type" model="ir.ui.view">
-        <field name="name">event.type.ticket.view.tree.inherit.sale</field>
+        <field name="name">event.type.ticket.view.tree.inherit.event.product</field>
         <field name="model">event.type.ticket</field>
         <field name="inherit_id" ref="event.event_type_ticket_view_tree_from_type"/>
         <field name="arch" type="xml">
@@ -22,7 +22,7 @@
     </record>
 
     <record id="event_type_ticket_view_form_from_type" model="ir.ui.view">
-        <field name="name">event.type.ticket.view.form.inherit.sale</field>
+        <field name="name">event.type.ticket.view.form.inherit.event.product</field>
         <field name="model">event.type.ticket</field>
         <field name="inherit_id" ref="event.event_type_ticket_view_form_from_type"/>
         <field name="arch" type="xml">
@@ -42,7 +42,7 @@
 
     <!-- EVENT.TICKET -->
     <record id="event_event_ticket_view_tree_from_event" model="ir.ui.view">
-        <field name="name">event.event.ticket.view.tree.from.event.inherit.sale</field>
+        <field name="name">event.event.ticket.view.tree.from.event.inherit.event.product</field>
         <field name="model">event.event.ticket</field>
         <field name="inherit_id" ref="event.event_event_ticket_view_tree_from_event"/>
         <field name="arch" type="xml">
@@ -61,14 +61,13 @@
                 />
             </field>
             <field name="description" position="after">
-                <field name="currency_id" column_invisible="True"/>
                 <field name="price" widget="monetary" options="{'currency_field': 'currency_id'}"/>
             </field>
         </field>
     </record>
 
     <record id="event_event_ticket_view_form_from_event" model="ir.ui.view">
-        <field name="name">event.event.ticket.view.form.from.event.inherit.sale</field>
+        <field name="name">event.event.ticket.view.form.from.event.inherit.event.product</field>
         <field name="model">event.event.ticket</field>
         <field name="inherit_id" ref="event.event_event_ticket_view_form_from_event"/>
         <field name="arch" type="xml">
@@ -87,7 +86,7 @@
     </record>
 
     <record id="event_event_ticket_view_kanban_from_event" model="ir.ui.view">
-        <field name="name">event.event.ticket.view.kanban.from.event</field>
+        <field name="name">event.event.ticket.view.kanban.from.event.product</field>
         <field name="model">event.event.ticket</field>
         <field name="inherit_id" ref="event.event_event_ticket_view_kanban_from_event"/>
         <field name="arch" type="xml">
@@ -105,7 +104,7 @@
     </record>
 
     <record id="event_event_ticket_form_view" model="ir.ui.view">
-        <field name="name">event.event.ticket.view.form.inherit.sale</field>
+        <field name="name">event.event.ticket.view.form.inherit.event.product</field>
         <field name="model">event.event.ticket</field>
         <field name="inherit_id" ref="event.event_event_ticket_form_view"/>
         <field name="arch" type="xml">
