+++ addons/event_sale/__manifest__.py
@@ -18,9 +18,8 @@ that product, you will be able to choose an existing event of that category and
 when you confirm your sales order it will automatically create a registration for
 this event.
 """,
-    'depends': ['event', 'sale_management'],
+    'depends': ['event_product', 'sale_management'],
     'data': [
-        'views/event_ticket_views.xml',
         'views/event_registration_views.xml',
         'views/event_views.xml',
         'views/product_template_views.xml',
@@ -36,7 +35,6 @@ this event.
     ],
     'demo': [
         'data/event_sale_demo.xml',
-        'data/event_demo.xml',  # needs event_sale_demo
         'data/event_registration_demo.xml',  # needs event_sale_demo
     ],
     'installable': True,
