+++ addons/event_sale/tests/common.py
@@ -4,38 +4,15 @@
 from datetime import datetime, timedelta
 
 from odoo import fields
-from odoo.addons.event.tests.common import EventCase
+from odoo.addons.event_product.tests.common import TestEventProductCommon
 from odoo.addons.sales_team.tests.common import TestSalesCommon
 
 
-class TestEventSaleCommon(EventCase, TestSalesCommon):
+class TestEventSaleCommon(TestEventProductCommon, TestSalesCommon):
 
     @classmethod
     def setUpClass(cls):
         super(TestEventSaleCommon, cls).setUpClass()
-
-        cls.event_product = cls.env['product.product'].create({
-            'name': 'Test Registration Product',
-            'description_sale': 'Mighty Description',
-            'list_price': 10,
-            'standard_price': 30.0,
-            'type': 'service',
-            'service_tracking': 'event',
-        })
-
-        cls.event_type_tickets = cls.env['event.type'].create({
-            'name': 'Update Type',
-            'has_seats_limitation': True,
-            'seats_max': 30,
-            'default_timezone': 'Europe/Paris',
-            'event_type_ticket_ids': [
-                (0, 0, {'name': 'First Ticket',
-                        'product_id': cls.event_product.id,
-                        'seats_max': 5,
-                       })
-            ],
-            'event_type_mail_ids': [],
-        })
 
         cls.event_0 = cls.env['event.event'].create({
             'name': 'TestEvent',
