+++ addons/sale/models/product_template.py
@@ -19,16 +19,6 @@ class ProductTemplate(models.Model):
         help="Manually set quantities on order: Invoice based on the manually entered quantity, without creating an analytic account.\n"
              "Timesheets on contract: Invoice based on the tracked hours on the related timesheet.\n"
              "Create a task and track hours: Create a task on the sales order validation and track the work hours.")
-    service_tracking = fields.Selection(selection=[
-            ('no', 'Nothing'),
-        ],
-        string="Create on Order",
-        default="no",
-        compute="_compute_service_tracking",
-        required=True,
-        store=True,
-        readonly=False,
-    )
     sale_line_warn = fields.Selection(
         WARNING_MESSAGE, string="Sales Order Line",
         help=WARNING_HELP, required=True, default="no-message")
@@ -101,7 +91,7 @@ class ProductTemplate(models.Model):
     def _prepare_service_tracking_tooltip(self):
         return ""
 
-    @api.depends('type', 'sale_ok')
+    @api.depends('sale_ok')
     def _compute_service_tracking(self):
         non_service_products = self.filtered(
             lambda pt: not pt.sale_ok or pt.type != 'service'
