+++ addons/test_event_full/data/event_type_data.xml
@@ -59,13 +59,13 @@
             (5, 0),
             (0, 0, {'description': 'Ticket1 Description',
                     'name': 'Ticket1',
-                    'product_id': ref('event_sale.product_product_event'),
+                    'product_id': ref('event_product.product_product_event'),
                     'seats_max': 10,
                    }
             ),
             (0, 0, {'description': 'Ticket2 Description',
                     'name': 'Ticket2',
-                    'product_id': ref('event_sale.product_product_event'),
+                    'product_id': ref('event_product.product_product_event'),
                     'price': 45,
                    }
             )]"/>
