+++ addons/website_event/security/ir.model.access.csv
@@ -17,13 +17,7 @@ access_website_event_menu_portal,website.event.menu,model_website_event_menu,bas
 access_website_event_menu_employee,website.event.menu,model_website_event_menu,base.group_user,1,0,0,0
 access_website_event_menu_user,website.event.menu.user,model_website_event_menu,event.group_event_user,1,1,1,1
 access_website_visitor_user,website.visitor.user,model_website_visitor,event.group_event_user,1,1,0,0
-access_event_question_public,event.question,model_event_question,base.group_public,1,0,0,0
-access_event_question_portal,event.question,model_event_question,base.group_portal,1,0,0,0
-access_event_question_employee,event.question,model_event_question,base.group_user,1,0,0,0
-access_event_question_user,event.question.user,model_event_question,event.group_event_user,1,1,1,1
-access_event_question_answer_public,event.question.answer,model_event_question_answer,base.group_public,1,0,0,0
-access_event_question_answer_portal,event.question.answer,model_event_question_answer,base.group_portal,1,0,0,0
-access_event_question_answer_employee,event.question.answer,model_event_question_answer,base.group_user,1,0,0,0
-access_event_question_answer_registration,event.question.answer.registration,model_event_question_answer,event.group_event_registration_desk,1,1,0,0
-access_event_question_answer_user,event.question.answer.user,model_event_question_answer,event.group_event_user,1,1,1,1
-access_event_registration_answer,event.registration.answer,model_event_registration_answer,event.group_event_registration_desk,1,1,1,1
+access_event_question_public,event.question,event.model_event_question,base.group_public,1,0,0,0
+access_event_question_portal,event.question,event.model_event_question,base.group_portal,1,0,0,0
+access_event_question_answer_public,event.question.answer,event.model_event_question_answer,base.group_public,1,0,0,0
+access_event_question_answer_portal,event.question.answer,event.model_event_question_answer,base.group_portal,1,0,0,0
